feat: better splitting
